+++ test/pipeline/pipeline.cmake
@@ -2,12 +2,7 @@
 #**********************
 # Gather Sources
 #**********************
-set(APP_SOURCES
-    ${CMAKE_CURRENT_LIST_DIR}/src/main.c
-    ${CMAKE_CURRENT_LIST_DIR}/src/xscope_fileio_task.c
-    ${CMAKE_CURRENT_LIST_DIR}/src/wav/wav_utils.c
-)
-
+file(GLOB_RECURSE APP_SOURCES ${CMAKE_CURRENT_LIST_DIR}/src/*.c )
 set(APP_INCLUDES
     ${CMAKE_CURRENT_LIST_DIR}/src
     ${CMAKE_CURRENT_LIST_DIR}/src/wav
@@ -25,30 +20,22 @@ endif()
 
 if(${TEST_PIPELINE} STREQUAL "FFVA_ADEC")
     message(STATUS "Building FFVA ADEC pipeline test")
-    set(APP_SOURCES
-        ${APP_SOURCES}
-        ${CMAKE_CURRENT_LIST_DIR}/src/ffva_adaptor.c
-    )
     set(AUDIO_PIPELINE_LIBRARY sln_voice::app::ffva::ap::adec)
     set(AUDIO_PIPELINE_INPUT_CHANNELS 4)
-    set(AUDIO_PIPELINE_USES_TILE_0 1)
-    set(AUDIO_PIPELINE_USES_TILE_1 1)
+    set(AUDIO_PIPELINE_INPUT_TILE_NO 1)
+    set(AUDIO_PIPELINE_OUTPUT_TILE_NO 0)
     set(TEST_PIPELINE_NAME test_pipeline_ffva_adec)
 elseif(${TEST_PIPELINE} STREQUAL "FFD")
     message(STATUS "Building FFD pipeline test")
     # The FFD pipeline needs other include paths set.  Gross!
-    set(APP_SOURCES
-        ${APP_SOURCES}
-        ${CMAKE_CURRENT_LIST_DIR}/src/ffd_adaptor.c
-    )
     set(APP_INCLUDES
         ${APP_INCLUDES}
         ${CMAKE_CURRENT_SOURCE_DIR}/examples/ffd/src
     )
     set(AUDIO_PIPELINE_LIBRARY sln_voice::app::ffd::ap)
     set(AUDIO_PIPELINE_INPUT_CHANNELS 2)
-    set(AUDIO_PIPELINE_USES_TILE_0 0)
-    set(AUDIO_PIPELINE_USES_TILE_1 1)
+    set(AUDIO_PIPELINE_INPUT_TILE_NO 1)
+    set(AUDIO_PIPELINE_OUTPUT_TILE_NO 1)
     set(TEST_PIPELINE_NAME test_pipeline_ffd)
 else()
     message(FATAL_ERROR "Unable to build ${TEST_PIPELINE} pipeline test")
@@ -75,8 +62,8 @@ set(APP_COMPILE_DEFINITIONS
     XSCOPE_HOST_IO_ENABLED=1
     XSCOPE_HOST_IO_TILE=0
     appconfAUDIO_PIPELINE_INPUT_CHANNELS=${AUDIO_PIPELINE_INPUT_CHANNELS}
-    appconfAUDIO_PIPELINE_USES_TILE_0=${AUDIO_PIPELINE_USES_TILE_0}
-    appconfAUDIO_PIPELINE_USES_TILE_1=${AUDIO_PIPELINE_USES_TILE_1}
+    appconfAUDIO_PIPELINE_INPUT_TILE_NO=${AUDIO_PIPELINE_INPUT_TILE_NO}
+    appconfAUDIO_PIPELINE_OUTPUT_TILE_NO=${AUDIO_PIPELINE_OUTPUT_TILE_NO}
 )
 
 set(APP_LINK_OPTIONS
